flan/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
paho-mqtt==1.6.1
requests==2.27.1
urllib3==1.26.9